+++ db/migrate/20260522120000_add_editable_board_id_to_users.rb
@@ -1,0 +1,6 @@
+class AddEditableBoardIdToUsers < ActiveRecord::Migration[7.1]
+  def change
+    add_column :users, :editable_board_id, :bigint
+    add_index :users, :editable_board_id
+  end
+end