+++ db/schema.rb
@@ -10,7 +10,7 @@
 #
 # It's strongly recommended that you check this file into your version control system.
 
-ActiveRecord::Schema[7.1].define(version: 2026_05_16_163820) do
+ActiveRecord::Schema[7.1].define(version: 2026_05_22_180000) do
   # These are extensions that must be enabled in order to support this database
   enable_extension "fuzzystrmatch"
   enable_extension "pg_trgm"
@@ -917,12 +917,15 @@ ActiveRecord::Schema[7.1].define(version: 2026_05_16_163820) do
     t.integer "plan_credits_balance", default: 0, null: false
     t.integer "topup_credits_balance", default: 0, null: false
     t.datetime "plan_credits_reset_at"
+    t.bigint "editable_board_id"
+    t.datetime "editable_board_id_set_at"
     t.index ["authentication_token"], name: "index_users_on_authentication_token", unique: true
     t.index ["child_lookup_key"], name: "index_users_on_child_lookup_key", unique: true
     t.index ["confirmation_token"], name: "index_users_on_confirmation_token", unique: true
     t.index ["current_team_id"], name: "index_users_on_current_team_id"
     t.index ["delete_account_token"], name: "index_users_on_delete_account_token", unique: true
     t.index ["deleted_at"], name: "index_users_on_deleted_at"
+    t.index ["editable_board_id"], name: "index_users_on_editable_board_id"
     t.index ["email"], name: "index_users_on_email", unique: true
     t.index ["invitation_token"], name: "index_users_on_invitation_token", unique: true
     t.index ["jti"], name: "index_users_on_jti", unique: true